+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Added extra precautions for when file loading fails and added backups of those files
 * Added a chunk border system that will show particles when close to an edge of a claimed chunk, closes #79
 * Added a config option to configure blocks that will ignore all the restrictions
+* Added an option to remove pvp between party members (Friendly Fire), closes #23
 
 # 1.0.14